TST: moved file test_concat.py to folder ./concat/ (#37243) (#37360)
